director-ui/pyproject.toml
[project]
name = "pdf-link-youtube-to-anything-tg-bot"
version = "0.1.0"
description = "A Telegram bot built with the telegram-bot-template framework"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Jack Ma", email = "hustlequeen@mail.ru"}
]
license = {text = "MIT"}
keywords = ["telegram", "bot", "framework", "ai", "openrouter", "openai", "extensible"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "aiohttp>=3.9.0",
    "alembic>=1.13.0",
    "asyncpg>=0.30.0",
    "click>=8.1.8",
    "openai>=1.75.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
    "pyyaml>=6.0.0",
    "sqlalchemy>=2.0.0",
    "pypdf>=4.0.0",
    "yt-dlp>=2024.1.0",
    "requests>=2.31.0",
    "aiofiles>=23.2.0",
    "crawl4ai>=0.3.0",
    "youtube-transcript-api>=1.2.2",
    "pytube>=15.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "python-socketio>=5.10.0",
    "websockets>=12.0",
    "python-json-logger>=2.0.7",
    "mediaempire-shared",
    "python-pptx>=1.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "coverage>=7.0.0",
    # "aiosqlite>=0.19.0", # No longer primary for tests if using Dockerized Postgres
    "testcontainers[postgres]>=3.7.0" # For Dockerized PostgreSQL in tests
]

[project.scripts]
pdf-link-youtube-to-anything-tg-bot = "main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 140
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session" # Added to address deprecation warning
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
]

[tool.coverage.run]
source = ["pdf_link_youtube_to_anything_tg_bot"]
omit = [
    "*/tests/*",
    "*/examples/*",
    "*/__pycache__/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.uv.sources]
mediaempire-shared = { path = "../mediaempire-shared", editable = true }

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.23.0",
]
